Extra fix for #6 Intersect works now
